+++ src/main/java/com/community/tools/service/GiveNewTaskService.java
@@ -1,7 +1,7 @@
 package com.community.tools.service;
 
-import com.community.tools.util.statemachine.Event;
-import com.community.tools.util.statemachine.State;
+import com.community.tools.util.statemachie.Event;
+import com.community.tools.util.statemachie.State;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.statemachine.StateMachine;
