+++ src/main/java/com/community/tools/service/github/AddMentorService.java
@@ -3,8 +3,8 @@ package com.community.tools.service.github;
 import com.community.tools.service.MessageService;
 import com.community.tools.service.StateMachineService;
 import com.community.tools.service.github.jpa.MentorsRepository;
-import com.community.tools.util.statemachine.Event;
-import com.community.tools.util.statemachine.State;
+import com.community.tools.util.statemachie.Event;
+import com.community.tools.util.statemachie.State;
 
 import com.github.seratch.jslack.api.methods.SlackApiException;
 import java.io.IOException;
