+++ src/main/java/com/community/tools/util/statemachie/Event.java
@@ -1,0 +1,7 @@
+package com.community.tools.util.statemachie;
+
+public enum Event {
+  ADD_GIT_NAME_AND_FIRST_TASK, GET_THE_FIRST_TASK, GET_THE_NEW_TASK, CHANGE_TASK, LAST_TASK,
+  LOGIN_CONFIRMATION, DID_NOT_PASS_VERIFICATION_GIT_LOGIN, QUESTION_FIRST, QUESTION_SECOND,
+  QUESTION_THIRD, CONSENT_TO_INFORMATION, SEND_ESTIMATE_TASK, CONFIRM_ESTIMATE;
+}
