+++ src/main/java/com/community/tools/util/statemachie/actions/transitions/questions/SecondQuestionActionTransition.java
@@ -1,0 +1,57 @@
+package com.community.tools.util.statemachie.actions.transitions.questions;
+
+import static com.community.tools.util.statemachie.Event.QUESTION_SECOND;
+import static com.community.tools.util.statemachie.State.FIRST_QUESTION;
+import static com.community.tools.util.statemachie.State.SECOND_QUESTION;
+
+import com.community.tools.model.User;
+import com.community.tools.service.MessageService;
+import com.community.tools.service.payload.QuestionPayload;
+import com.community.tools.util.statemachie.Event;
+import com.community.tools.util.statemachie.State;
+import com.community.tools.util.statemachie.actions.Transition;
+import com.community.tools.util.statemachie.jpa.StateMachineRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.statemachine.StateContext;
+import org.springframework.statemachine.action.Action;
+import org.springframework.statemachine.annotation.WithStateMachine;
+import org.springframework.statemachine.config.builders.StateMachineTransitionConfigurer;
+
+@WithStateMachine
+public class SecondQuestionActionTransition implements Transition {
+
+  @Value("${secondQuestion}")
+  private String secondQuestion;
+
+  @Autowired
+  private MessageService messageService;
+
+  @Autowired
+  private Action<State, Event> errorAction;
+
+  @Autowired
+  private StateMachineRepository stateMachineRepository;
+
+  @Override
+  public void configure(
+      StateMachineTransitionConfigurer<State, Event> transitions) throws Exception {
+    transitions
+        .withExternal()
+        .source(FIRST_QUESTION)
+        .target(SECOND_QUESTION)
+        .event(QUESTION_SECOND)
+        .action(this, errorAction);
+  }
+
+  @Override
+  public void execute(StateContext<State, Event> stateContext) {
+    QuestionPayload payloadFirstAnswer = (QuestionPayload) stateContext.getExtendedState()
+        .getVariables().get("dataPayload");
+    String id = payloadFirstAnswer.getUser();
+    User stateEntity = stateMachineRepository.findByUserID(id).get();
+    stateEntity.setFirstAnswerAboutRules(payloadFirstAnswer.getAnswer());
+    stateMachineRepository.save(stateEntity);
+    messageService.sendBlocksMessage(messageService.getUserById(id), secondQuestion);
+  }
+}
