+++ src/main/java/com/community/tools/util/statemachie/actions/transitions/questions/ThirdQuestionActionTransition.java
@@ -1,0 +1,57 @@
+package com.community.tools.util.statemachie.actions.transitions.questions;
+
+import static com.community.tools.util.statemachie.Event.QUESTION_THIRD;
+import static com.community.tools.util.statemachie.State.SECOND_QUESTION;
+import static com.community.tools.util.statemachie.State.THIRD_QUESTION;
+
+import com.community.tools.model.User;
+import com.community.tools.service.MessageService;
+import com.community.tools.service.payload.QuestionPayload;
+import com.community.tools.util.statemachie.Event;
+import com.community.tools.util.statemachie.State;
+import com.community.tools.util.statemachie.actions.Transition;
+import com.community.tools.util.statemachie.jpa.StateMachineRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.statemachine.StateContext;
+import org.springframework.statemachine.action.Action;
+import org.springframework.statemachine.annotation.WithStateMachine;
+import org.springframework.statemachine.config.builders.StateMachineTransitionConfigurer;
+
+@WithStateMachine
+public class ThirdQuestionActionTransition implements Transition {
+
+  @Value("${thirdQuestion}")
+  private String thirdQuestion;
+
+  @Autowired
+  private MessageService messageService;
+
+  @Autowired
+  private Action<State, Event> errorAction;
+
+  @Autowired
+  private StateMachineRepository stateMachineRepository;
+
+  @Override
+  public void configure(
+      StateMachineTransitionConfigurer<State, Event> transitions) throws Exception {
+    transitions
+        .withExternal()
+        .source(SECOND_QUESTION)
+        .target(THIRD_QUESTION)
+        .event(QUESTION_THIRD)
+        .action(this, errorAction);
+  }
+
+  @Override
+  public void execute(StateContext<State, Event> stateContext) {
+    QuestionPayload payloadSecondAnswer = (QuestionPayload) stateContext.getExtendedState()
+        .getVariables().get("dataPayload");
+    String id = payloadSecondAnswer.getUser();
+    User stateEntity = stateMachineRepository.findByUserID(id).get();
+    stateEntity.setSecondAnswerAboutRules(payloadSecondAnswer.getAnswer());
+    stateMachineRepository.save(stateEntity);
+    messageService.sendBlocksMessage(messageService.getUserById(id), thirdQuestion);
+  }
+}
