+++ src/main/java/com/community/tools/util/statemachie/actions/transitions/tasks/GetTheFirstTaskActionTransition.java
@@ -1,0 +1,49 @@
+package com.community.tools.util.statemachie.actions.transitions.tasks;
+
+import static com.community.tools.util.statemachie.Event.GET_THE_FIRST_TASK;
+import static com.community.tools.util.statemachie.State.ADDED_GIT;
+import static com.community.tools.util.statemachie.State.ESTIMATE_THE_TASK;
+
+import com.community.tools.service.MessageService;
+import com.community.tools.service.payload.SinglePayload;
+import com.community.tools.util.statemachie.Event;
+import com.community.tools.util.statemachie.State;
+import com.community.tools.util.statemachie.actions.Transition;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.statemachine.StateContext;
+import org.springframework.statemachine.action.Action;
+import org.springframework.statemachine.annotation.WithStateMachine;
+import org.springframework.statemachine.config.builders.StateMachineTransitionConfigurer;
+
+@WithStateMachine
+public class GetTheFirstTaskActionTransition implements Transition {
+
+  @Value("${getFirstTask}")
+  private String getFirstTask;
+
+  @Autowired
+  private MessageService messageService;
+
+  @Autowired
+  private Action<State, Event> errorAction;
+
+  @Override
+  public void configure(
+      StateMachineTransitionConfigurer<State, Event> transitions) throws Exception {
+    transitions
+        .withExternal()
+        .source(ADDED_GIT)
+        .target(ESTIMATE_THE_TASK)
+        .event(GET_THE_FIRST_TASK)
+        .action(this, errorAction);
+  }
+
+  @Override
+  public void execute(StateContext<State, Event> stateContext) {
+    SinglePayload payload = (SinglePayload) stateContext.getExtendedState().getVariables()
+        .get("dataPayload");
+    String user = payload.getId();
+    messageService.sendBlocksMessage(messageService.getUserById(user), getFirstTask);
+  }
+}
