--- src/main/java/com/community/tools/util/statemachine/State.java
@@ -1,8 +1,0 @@
-package com.community.tools.util.statemachine;
-
-
-public enum State {
-  NEW_USER, AGREED_LICENSE, ADDED_GIT, GOT_THE_TASK, CHECK_FOR_NEW_TASK,
-  CONGRATS_LAST_TASK, CHECK_LOGIN, FIRST_QUESTION, SECOND_QUESTION, THIRD_QUESTION,
-  ESTIMATE_THE_TASK, CHECK_ESTIMATE
-}
