+++ src/test/java/com/community/tools/service/PointsTaskServiceTest.java
@@ -7,9 +7,9 @@ import static org.mockito.Mockito.when;
 import com.community.tools.model.Mentors;
 import com.community.tools.model.User;
 import com.community.tools.service.github.jpa.MentorsRepository;
-import com.community.tools.util.statemachine.Event;
-import com.community.tools.util.statemachine.State;
-import com.community.tools.util.statemachine.jpa.StateMachineRepository;
+import com.community.tools.util.statemachie.Event;
+import com.community.tools.util.statemachie.State;
+import com.community.tools.util.statemachie.jpa.StateMachineRepository;
 
 import java.util.Arrays;
 import java.util.HashMap;
